+++ webknossos/webknossos/dataset/layer/abstract_layer.py
@@ -23,6 +23,7 @@ from webknossos.geometry import NDBoundingBox, NormalizedBoundingBox
 from webknossos.geometry.mag import Mag, MagLike
 
 from ...utils import warn_deprecated
+from .export import LayerExport
 from .view import ArrayException, MagView
 
 if TYPE_CHECKING:
@@ -56,6 +57,7 @@ class AbstractLayer:
     ) -> None:
         self._dataset = dataset
         self._apply_properties(properties, read_only)
+        self._export = LayerExport(self)
 
     def _apply_properties(self, properties: LayerProperties, read_only: bool) -> None:
         # It is possible that the properties on disk do not contain the number of channels.
@@ -349,6 +351,16 @@ class AbstractLayer:
         """
 
         return self._name
+
+    @property
+    def export(self) -> LayerExport:
+        """Export this layer's data to common bioimaging file formats.
+
+        Returns:
+            LayerExport: Accessor object with `as_ozx`, `as_tiff_stack` and
+                `as_ome_tiff` export methods.
+        """
+        return self._export
 
     def get_mag(self, mag: MagLike) -> MagView:
         """Gets the MagView for the specified magnification level.